+++ src/amocrm/dto/__init__.py
@@ -1,0 +1,5 @@
+from amocrm.dto.customer import AmoCRMCustomer
+
+__all__ = [
+    "AmoCRMCustomer",
+]
