+++ src/amocrm/dto/base.py
