--- src/amocrm/entities/__init__.py
@@ -1,5 +1,0 @@
-from amocrm.entities.customer import AmoCRMCustomer
-
-__all__ = [
-    "AmoCRMCustomer",
-]
