+++ src/amocrm/tests/dto/conftest.py
@@ -3,7 +3,7 @@ import pytest
 
 @pytest.fixture(autouse=True)
 def _mock_contact_field_id(mocker):
-    mocker.patch("amocrm.entities.customer.get_contact_field_id", return_value=2235143)
+    mocker.patch("amocrm.dto.customer.get_contact_field_id", return_value=2235143)
 
 
 @pytest.fixture
